Changes to be incorporated:
1. NewRequest
- in NewRequest tab the
input box for AWR No. to be a SearchableComboBox

and this combobox is to refer a csv file
"AWR_Issuance_Cumulative_Master.csv" (first column "File Name")

this particular sheet will be updated by a person so that the user has current and correct awr no. (file-name)

it will be at a loaction that is to be set

so that user can search and select the awr no. in the combox

- also need to bring back the ability for QtyRequired and QtyIssued issued , this is enabled only when the user selects RM as awr type

as RM-specific have sublots therefore they need multiple prints of the same file (awr)

otherwise all others are one request - one document by default